Implementing BookAuditService and calling it from corresponding aspect
