+++ infrastructure/src/main/java/com/ros/aop/audit_service/BookAuditService.java
@@ -1,0 +1,6 @@
+package com.ros.aop.audit_service;
+
+public interface BookAuditService {
+    void logAddBookAfterReturning();
+    void logAddBookAfterThrowing();
+}
